Update findStateByStatus as per sharding schema changes
Updated the SQL query as per new schema changes for sharding.
This query works on a 'join' between the States and StateMachines
table. It retrieves all states having one of the given statuses for
a particular state machine name and the state machine creation time in
the given range fromTime and toTime.