Update findErroredStates API
findErroredStates API used to work earlier based on StateMachineId
which was an integer earlier. However, as now it is changed to String,
findErroredStates API cannot work on a range on StateMachineIds, hence
changing it to a time range on StateMachine's creation time.

Note: Updating unit test "testGetErroredStates". Will push it soon.