ClientElb Persistence layer with in-memory cache
This layer introduces an in-memory cache to store ClientElb entries.
The cache is maintained in this Singleton instance object of
ClientElbPersistenceService. Cache operations viz. refresh of cache in
case of cache miss is based on LRU policy for which another map
LastAccessed is used. ConcurrentHashMap used to build the cache ensures
concurrency in writes/updates/delete but reads.
A layer of synchronized block (on singleton instance object of
ClientElbPersistenceService) is in place which ensures thread safe execution
occuring on the cache. Cache updates/deletes it's entries if database
update/delete happens after a double check before actually updating/deleting
entries from the cache.
Cache size is configured to 50 in yml file.