Remove double check
Earlier this double check was added because the thought was to refresh
cache by running a scheduled job every minute. Since now we are refreshing
cache only on update/delete within the family of synchronized block on
same singleton object instance of this class,
we don't need it anymore, hence removed.