Altered clientElbId default name
Changed clientElbId name to something logical like 'defaultElbId'
in FluxClientConfiguration and all other usages in tests, config files.

Made ClientElbId column in StateMachines table as NOT NULL and default value
set to 'defaultElbId' which will be set to all existing StateMachines
when Alter table query is applied. With this, having an entry with
'defaultElbId' in ClientElb table will route the SM executions correctly.

All tests passing.