[ExecutionVersionedWorkflows]
Added TraversalPathMapping data table. This table will store
for a given state, list of dependent states and events coming
in it's BFS traversal.
This will be populated once when state machine instance is created and
will be used when ReplayEvent is triggered to retrieve list of states
and events that needs to execute with new executionVersion.

All tests passing.