[ExecutionVersionedWorkflows]
Added an integration test for ReplayEvent trigger.

Replayable state are redrived by redriver, thread sleep in between
is set to 8 secs for each RE trigger because redriver batch read
interval is 2.5 secs and further sleep of 5.5 secs is for all states
to get completed (7 states, each with timeout of 0.5 secs).

Assertions are being done on completion status of states in replayable
state's traversal path.